+++ pdns/dnsdistdist/dnsdist-udp.cc
@@ -22,10 +22,13 @@
 
 #include "dnsdist.hh"
 #include "dnsdist-udp.hh"
+#include "dnsdist-dnscrypt.hh"
 #include "dnsdist-metrics.hh"
 #include "dnsparser.hh"
 #include "dnsdist-dnsparser.hh"
 #include "dnsdist-ecs.hh"
+#include "dnsdist-proxy-protocol.hh"
+#include "threadname.hh"
 #include "dolog.hh"
 #include "dnsdist-configuration.hh"
 
@@ -229,6 +232,488 @@ void handleResponseForUDPClient(InternalQueryState& ids, PacketBuffer& response,
   }
   else {
     handleResponseSent(ids, 0., dnsResponse.ids.origRemote, ComboAddress(), response.size(), cleartextDH, dnsdist::Protocol::DoUDP, false);
+  }
+}
+
+#if !defined(DISABLE_RECVMMSG) && defined(HAVE_RECVMMSG) && defined(HAVE_SENDMMSG) && defined(MSG_WAITFORONE)
+void queueResponse(const PacketBuffer& response, const ComboAddress& dest, const ComboAddress& remote, struct mmsghdr& outMsg, struct iovec* iov, cmsgbuf_aligned* cbuf)
+{
+  outMsg.msg_len = 0;
+  // NOLINTNEXTLINE(cppcoreguidelines-pro-type-const-cast,cppcoreguidelines-pro-type-reinterpret-cast): API
+  fillMSGHdr(&outMsg.msg_hdr, iov, nullptr, 0, const_cast<char*>(reinterpret_cast<const char*>(&response.at(0))), response.size(), const_cast<ComboAddress*>(&remote));
+
+  if (dest.sin4.sin_family == 0) {
+    outMsg.msg_hdr.msg_control = nullptr;
+  }
+  else {
+    addCMsgSrcAddr(&outMsg.msg_hdr, cbuf, &dest, 0);
+  }
+}
+#endif
+
+bool isUDPQueryAcceptable(ClientState& clientState, const struct msghdr* msgh, const ComboAddress& remote, ComboAddress& dest, bool& expectProxyProtocol)
+{
+  if ((msgh->msg_flags & MSG_TRUNC) != 0) {
+    /* message was too large for our buffer */
+    VERBOSESLOG(infolog("Dropping message too large for our buffer"),
+                dnsdist::logging::getTopLogger("udp-query")->info(Logr::Info, "Dropping query from client that is too large for our buffer", "client.address", Logging::Loggable(remote), "destination.address", Logging::Loggable(dest), "frontend.address", Logging::Loggable(clientState.local)));
+    ++dnsdist::metrics::g_stats.nonCompliantQueries;
+    ++clientState.nonCompliantQueries;
+    return false;
+  }
+
+  expectProxyProtocol = clientState.d_enableProxyProtocol && expectProxyProtocolFrom(remote);
+  if (!dnsdist::configuration::getCurrentRuntimeConfiguration().d_ACL.match(remote) && !expectProxyProtocol) {
+    VERBOSESLOG(infolog("Query from %s dropped because of ACL", remote.toStringWithPort()),
+                dnsdist::logging::getTopLogger("udp-query")->info(Logr::Info, "Query dropped because of ACL", "source.address", Logging::Loggable(dest)));
+    ++dnsdist::metrics::g_stats.aclDrops;
+    return false;
+  }
+
+  if (HarvestDestinationAddress(msgh, &dest)) {
+    /* so it turns out that sometimes the kernel lies to us:
+       the address is set to 0.0.0.0:0 which makes our sendfromto() use
+       the wrong address. In that case it's better to let the kernel
+       do the work by itself and use sendto() instead.
+       This is indicated by setting the family to 0 which is acted upon
+       in sendUDPResponse() and DelayedPacket::().
+    */
+    static const ComboAddress bogusV4("0.0.0.0:0");
+    static const ComboAddress bogusV6("[::]:0");
+    if ((dest.sin4.sin_family == AF_INET && dest == bogusV4) || (dest.sin4.sin_family == AF_INET6 && dest == bogusV6)) {
+      dest.sin4.sin_family = 0;
+    }
+    else {
+      /* we don't get the port, only the address */
+      dest.sin4.sin_port = clientState.local.sin4.sin_port;
+    }
+  }
+  else {
+    dest.sin4.sin_family = 0;
+  }
+
+  ++clientState.queries;
+  ++dnsdist::metrics::g_stats.queries;
+
+  return true;
+}
+
+void processUDPQuery(ClientState& clientState, const struct msghdr* msgh, const ComboAddress& remote, ComboAddress& dest, PacketBuffer& query, std::vector<mmsghdr>* responsesVect, unsigned int* queuedResponses, struct iovec* respIOV, cmsgbuf_aligned* respCBuf)
+{
+  assert(responsesVect == nullptr || (queuedResponses != nullptr && respIOV != nullptr && respCBuf != nullptr));
+  uint16_t queryId = 0;
+  InternalQueryState ids;
+
+  auto closer = ids.getCloser(__func__, SpanID{}); // NOLINT(cppcoreguidelines-pro-bounds-array-to-pointer-decay)
+
+  ids.cs = &clientState;
+  ids.origRemote = remote;
+  ids.hopRemote = remote;
+  ids.protocol = dnsdist::Protocol::DoUDP;
+
+  try {
+    bool expectProxyProtocol = false;
+    if (!isUDPQueryAcceptable(clientState, msgh, remote, dest, expectProxyProtocol)) {
+      return;
+    }
+    /* dest might have been updated, if we managed to harvest the destination address */
+    if (dest.sin4.sin_family != 0) {
+      ids.origDest = dest;
+      ids.hopLocal = dest;
+    }
+    else {
+      /* if we have not been able to harvest the destination address,
+         we do NOT want to update dest or hopLocal, to let the kernel
+         pick the less terrible option, but we want to update origDest
+         which is used by rules and actions to at least the correct
+         address family */
+      ids.origDest = clientState.local;
+      ids.hopLocal.sin4.sin_family = 0;
+    }
+
+    std::vector<ProxyProtocolValue> proxyProtocolValues;
+    if (expectProxyProtocol && !handleProxyProtocol(remote, false, dnsdist::configuration::getCurrentRuntimeConfiguration().d_ACL, query, ids.origRemote, ids.origDest, proxyProtocolValues)) {
+      return;
+    }
+
+    ids.queryRealTime.start();
+
+    auto dnsCryptResponse = dnsdist::dnscrypt::checkDNSCryptQuery(clientState, query, ids.dnsCryptQuery, ids.queryRealTime.d_start.tv_sec, false);
+    if (dnsCryptResponse) {
+      if (!clientState.muted) {
+        sendUDPResponse(clientState.udpFD, query, 0, dest, remote);
+      }
+      return;
+    }
+
+    {
+      /* this pointer will be invalidated the second the buffer is resized, don't hold onto it! */
+      const dnsheader_aligned dnsHeader(query.data());
+      queryId = ntohs(dnsHeader->id);
+
+      if (!checkQueryHeaders(*dnsHeader, clientState)) {
+        return;
+      }
+
+      if (dnsHeader->qdcount == 0) {
+        dnsdist::PacketMangling::editDNSHeaderFromPacket(query, [](dnsheader& header) {
+          header.rcode = RCode::NotImp;
+          header.qr = true;
+          return true;
+        });
+
+        if (!clientState.muted) {
+          sendUDPResponse(clientState.udpFD, query, 0, dest, remote);
+        }
+        return;
+      }
+    }
+
+    // NOLINTNEXTLINE(cppcoreguidelines-pro-type-reinterpret-cast)
+    ids.qname = DNSName(reinterpret_cast<const char*>(query.data()), query.size(), sizeof(dnsheader), false, &ids.qtype, &ids.qclass);
+    if (ids.dnsCryptQuery) {
+      ids.protocol = dnsdist::Protocol::DNSCryptUDP;
+    }
+    DNSQuestion dnsQuestion(ids, query);
+    const uint16_t* flags = getFlagsFromDNSHeader(dnsQuestion.getHeader().get());
+    ids.origFlags = *flags;
+
+    if (!proxyProtocolValues.empty()) {
+      dnsQuestion.proxyProtocolValues = make_unique<std::vector<ProxyProtocolValue>>(std::move(proxyProtocolValues));
+    }
+
+    // save UDP payload size from origin query
+    uint16_t udpPayloadSize = 0;
+    uint16_t zValue = 0;
+    // NOLINTNEXTLINE(cppcoreguidelines-pro-type-reinterpret-cast)
+    getEDNSUDPPayloadSizeAndZ(reinterpret_cast<const char*>(query.data()), query.size(), &udpPayloadSize, &zValue);
+    if (!ids.dnssecOK) {
+      ids.dnssecOK = (zValue & EDNS_HEADER_FLAG_DO) != 0;
+    }
+    if (udpPayloadSize < 512) {
+      udpPayloadSize = 512;
+    }
+
+    std::shared_ptr<DownstreamState> backend{nullptr};
+    auto result = processQuery(dnsQuestion, backend);
+
+    if (result == ProcessQueryResult::Drop || result == ProcessQueryResult::Asynchronous) {
+      return;
+    }
+
+    // the buffer might have been invalidated by now (resized)
+    const auto dnsHeader = dnsQuestion.getHeader();
+    if (result == ProcessQueryResult::SendAnswer) {
+      /* ensure payload size is not exceeded */
+      dnsdist::udp::handleResponseTC4UDPClient(dnsQuestion, udpPayloadSize, query);
+#ifndef DISABLE_RECVMMSG
+#if defined(HAVE_RECVMMSG) && defined(HAVE_SENDMMSG) && defined(MSG_WAITFORONE)
+      if (dnsQuestion.ids.delayMsec == 0 && responsesVect != nullptr) {
+        queueResponse(query, dest, remote, (*responsesVect)[*queuedResponses], respIOV, respCBuf);
+        (*queuedResponses)++;
+        handleResponseSent(std::move(dnsQuestion.ids.qname), dnsQuestion.ids.qtype, 0., remote, ComboAddress(), query.size(), *dnsHeader, dnsdist::Protocol::DoUDP, dnsdist::Protocol::DoUDP, false);
+        return;
+      }
+#endif /* defined(HAVE_RECVMMSG) && defined(HAVE_SENDMMSG) && defined(MSG_WAITFORONE) */
+#endif /* DISABLE_RECVMMSG */
+      /* we use dest, always, because we don't want to use the listening address to send a response since it could be 0.0.0.0 */
+      if (!clientState.muted) {
+        sendUDPResponse(clientState.udpFD, query, dnsQuestion.ids.delayMsec, dest, remote);
+      }
+
+      handleResponseSent(std::move(dnsQuestion.ids.qname), dnsQuestion.ids.qtype, 0., remote, ComboAddress(), query.size(), *dnsHeader, dnsdist::Protocol::DoUDP, dnsdist::Protocol::DoUDP, false);
+      return;
+    }
+
+    if (result != ProcessQueryResult::PassToBackend || backend == nullptr) {
+      return;
+    }
+
+    if (backend->isTCPOnly()) {
+      std::string proxyProtocolPayload;
+      /* we need to do this _before_ creating the cross protocol query because
+         after that the buffer will have been moved */
+      if (backend->d_config.useProxyProtocol) {
+        proxyProtocolPayload = getProxyProtocolPayload(dnsQuestion);
+      }
+
+      ids.origID = dnsHeader->id;
+      auto cpq = std::make_unique<dnsdist::udp::UDPCrossProtocolQuery>(std::move(query), std::move(ids), backend);
+      cpq->query.d_proxyProtocolPayload = std::move(proxyProtocolPayload);
+
+      backend->passCrossProtocolQuery(std::move(cpq));
+      return;
+    }
+
+    assignOutgoingUDPQueryToBackend(backend, dnsHeader->id, dnsQuestion, query);
+  }
+  catch (const std::exception& e) {
+    VERBOSESLOG(infolog("Got an error in UDP question thread while parsing a query from %s, id %d: %s", ids.origRemote.toStringWithPort(), queryId, e.what()),
+                dnsdist::logging::getTopLogger("udp-frontend")->error(Logr::Info, e.what(), "Got an error in UDP question thread while parsing a query", "source.address", Logging::Loggable(ids.origRemote), "dns.question.id", Logging::Loggable(queryId)));
+  }
+}
+
+size_t getInitialUDPPacketBufferSize(bool expectProxyProtocol)
+{
+  static_assert(dnsdist::configuration::s_udpIncomingBufferSize <= dnsdist::udp::s_initialUDPPacketBufferSize, "The incoming buffer size should not be larger than s_initialUDPPacketBufferSize");
+
+  const auto& runtimeConfig = dnsdist::configuration::getCurrentRuntimeConfiguration();
+  if (!expectProxyProtocol || runtimeConfig.d_proxyProtocolACL.empty()) {
+    return s_initialUDPPacketBufferSize;
+  }
+
+  return s_initialUDPPacketBufferSize + runtimeConfig.d_proxyProtocolMaximumSize;
+}
+
+size_t getMaximumIncomingPacketSize(const ClientState& clientState)
+{
+  if (clientState.dnscryptCtx) {
+    return getInitialUDPPacketBufferSize(clientState.d_enableProxyProtocol);
+  }
+
+  const auto& runtimeConfig = dnsdist::configuration::getCurrentRuntimeConfiguration();
+  if (!clientState.d_enableProxyProtocol || runtimeConfig.d_proxyProtocolACL.empty()) {
+    return dnsdist::configuration::s_udpIncomingBufferSize;
+  }
+
+  return dnsdist::configuration::s_udpIncomingBufferSize + runtimeConfig.d_proxyProtocolMaximumSize;
+}
+
+#ifndef DISABLE_RECVMMSG
+#if defined(HAVE_RECVMMSG) && defined(HAVE_SENDMMSG) && defined(MSG_WAITFORONE)
+void MultipleMessagesUDPClientThread(ClientState* clientState)
+{
+  struct MMReceiver
+  {
+    PacketBuffer packet;
+    ComboAddress remote;
+    ComboAddress dest;
+    iovec iov{};
+    /* used by HarvestDestinationAddress */
+    cmsgbuf_aligned cbuf{};
+  };
+  const size_t vectSize = dnsdist::configuration::getImmutableConfiguration().d_udpVectorSize;
+
+  if (vectSize > std::numeric_limits<uint16_t>::max()) {
+    throw std::runtime_error("The value of setUDPMultipleMessagesVectorSize is too high, the maximum value is " + std::to_string(std::numeric_limits<uint16_t>::max()));
+  }
+
+  auto recvData = std::vector<MMReceiver>(vectSize);
+  auto msgVec = std::vector<mmsghdr>(vectSize);
+  auto outMsgVec = std::vector<mmsghdr>(vectSize);
+
+  /* the actual buffer is larger because:
+     - we may have to add EDNS and/or ECS
+     - we use it for self-generated responses (from rule or cache)
+     but we only accept incoming payloads up to that size
+  */
+  const size_t initialBufferSize = getInitialUDPPacketBufferSize(clientState->d_enableProxyProtocol);
+  const size_t maxIncomingPacketSize = getMaximumIncomingPacketSize(*clientState);
+
+  /* initialize the structures needed to receive our messages */
+  for (size_t idx = 0; idx < vectSize; idx++) {
+    recvData[idx].remote.sin4.sin_family = clientState->local.sin4.sin_family;
+    recvData[idx].packet.resize(initialBufferSize);
+    // NOLINTNEXTLINE(cppcoreguidelines-pro-type-reinterpret-cast)
+    fillMSGHdr(&msgVec[idx].msg_hdr, &recvData[idx].iov, &recvData[idx].cbuf, sizeof(recvData[idx].cbuf), reinterpret_cast<char*>(recvData[idx].packet.data()), maxIncomingPacketSize, &recvData[idx].remote);
+  }
+
+  int msgsGot = static_cast<int>(vectSize);
+  /* go now */
+  for (;;) {
+
+    /* reset the IO vector, since it's also used to send the vector of responses
+       to avoid having to copy the data around.
+       No need to reset the parts that have not been used, though. */
+    for (int idx = 0; idx < msgsGot; idx++) {
+      auto& slot = recvData[idx];
+      /* only resize if the buffer is actually smaller than expected */
+      if (slot.packet.size() < initialBufferSize) {
+        slot.packet.resize(initialBufferSize);
+      }
+      /* but we need to set the IOv pointer and size
+         anyway, because if we resized it the pointer might
+         now be invalid */
+      slot.iov.iov_base = &slot.packet.at(0);
+      slot.iov.iov_len = slot.packet.size();
+    }
+
+    /* block until we have at least one message ready, but return
+       as many as possible to save the syscall costs */
+    msgsGot = recvmmsg(clientState->udpFD, msgVec.data(), vectSize, MSG_WAITFORONE, nullptr);
+    if (msgsGot <= 0) {
+      int savederrno = errno;
+      VERBOSESLOG(infolog("Getting UDP messages via recvmmsg() failed with: %s", stringerror(savederrno)),
+                  dnsdist::logging::getTopLogger("udp-recvmmsg-frontend")->error(Logr::Info, savederrno, "Getting UDP messages via recvmmsg failed", "frontend.address", Logging::Loggable(clientState->local)));
+      msgsGot = 0;
+      continue;
+    }
+
+    unsigned int msgsToSend = 0;
+
+    /* process the received messages */
+    for (int msgIdx = 0; msgIdx < msgsGot; msgIdx++) {
+      auto& msg = msgVec[msgIdx];
+      const struct msghdr* msgh = &msg.msg_hdr;
+      unsigned int got = msg.msg_len;
+      const ComboAddress& remote = recvData[msgIdx].remote;
+
+      if (static_cast<size_t>(got) < sizeof(struct dnsheader)) {
+        ++dnsdist::metrics::g_stats.nonCompliantQueries;
+        ++clientState->nonCompliantQueries;
+        continue;
+      }
+
+      if ((msgh->msg_flags & MSG_TRUNC) != 0) {
+        /* message was too large for our buffer */
+        ++dnsdist::metrics::g_stats.nonCompliantQueries;
+        ++clientState->nonCompliantQueries;
+        continue;
+      }
+
+      auto& data = recvData[msgIdx];
+      data.packet.resize(got);
+      dnsdist::configuration::refreshLocalRuntimeConfiguration();
+      processUDPQuery(*clientState, msgh, remote, data.dest, data.packet, &outMsgVec, &msgsToSend, &data.iov, &data.cbuf);
+    }
+
+    /* immediate (not delayed or sent to a backend) responses (mostly from a rule, dynamic block
+       or the cache) can be sent in batch too */
+
+    if (msgsToSend > 0 && msgsToSend <= static_cast<unsigned int>(msgsGot)) {
+      int sent = sendmmsg(clientState->udpFD, outMsgVec.data(), msgsToSend, 0);
+
+      if (sent < 0 || static_cast<unsigned int>(sent) != msgsToSend) {
+        int savederrno = errno;
+        VERBOSESLOG(infolog("Error sending responses with sendmmsg() (%d on %u): %s", sent, msgsToSend, stringerror(savederrno)),
+                    dnsdist::logging::getTopLogger("udp-sendmmsg-frontend")->error(Logr::Info, savederrno, "Error sending responses with sendmmsg()", "address", Logging::Loggable(clientState->local), "dnsdist.sendmmsg.messages_sent", Logging::Loggable(sent), "dnsdist.sendmmsg.messages_to_send", Logging::Loggable(msgsToSend)));
+      }
+    }
+  }
+}
+#endif /* defined(HAVE_RECVMMSG) && defined(HAVE_SENDMMSG) && defined(MSG_WAITFORONE) */
+#endif /* DISABLE_RECVMMSG */
+
+// listens to incoming queries, sends out to downstream servers, noting the intended return path
+void udpClientThread(std::vector<ClientState*> states)
+{
+  try {
+    setThreadName("dnsdist/udpClie");
+#ifndef DISABLE_RECVMMSG
+#if defined(HAVE_RECVMMSG) && defined(HAVE_SENDMMSG) && defined(MSG_WAITFORONE)
+    if (dnsdist::configuration::getImmutableConfiguration().d_udpVectorSize > 1) {
+      MultipleMessagesUDPClientThread(states.at(0));
+    }
+    else
+#endif /* defined(HAVE_RECVMMSG) && defined(HAVE_SENDMMSG) && defined(MSG_WAITFORONE) */
+#endif /* DISABLE_RECVMMSG */
+    {
+      /* the actual buffer is larger because:
+         - we may have to add EDNS and/or ECS
+         - we use it for self-generated responses (from rule or cache)
+         but we only accept incoming payloads up to that size
+      */
+      struct UDPStateParam
+      {
+        ClientState* cs{nullptr};
+        size_t maxIncomingPacketSize{0};
+        int socket{-1};
+      };
+      const size_t initialBufferSize = getInitialUDPPacketBufferSize(true);
+      PacketBuffer packet(initialBufferSize);
+
+      msghdr msgh{};
+      iovec iov{};
+      ComboAddress remote;
+      ComboAddress dest;
+
+      auto handleOnePacket = [&packet, &iov, &msgh, &remote, &dest, initialBufferSize](const UDPStateParam& param) {
+        packet.resize(initialBufferSize);
+        iov.iov_base = &packet.at(0);
+        iov.iov_len = packet.size();
+
+        ssize_t got = recvmsg(param.socket, &msgh, 0);
+
+        if (got < 0 || static_cast<size_t>(got) < sizeof(struct dnsheader)) {
+          ++dnsdist::metrics::g_stats.nonCompliantQueries;
+          ++param.cs->nonCompliantQueries;
+          return;
+        }
+
+        packet.resize(static_cast<size_t>(got));
+
+        dnsdist::configuration::refreshLocalRuntimeConfiguration();
+        processUDPQuery(*param.cs, &msgh, remote, dest, packet, nullptr, nullptr, nullptr, nullptr);
+      };
+
+      std::vector<UDPStateParam> params;
+      for (auto& state : states) {
+        const size_t maxIncomingPacketSize = getMaximumIncomingPacketSize(*state);
+        params.emplace_back(UDPStateParam{state, maxIncomingPacketSize, state->udpFD});
+      }
+
+      if (params.size() == 1) {
+        const auto& param = params.at(0);
+        remote.sin4.sin_family = param.cs->local.sin4.sin_family;
+        /* used by HarvestDestinationAddress */
+        cmsgbuf_aligned cbuf;
+        // NOLINTNEXTLINE(cppcoreguidelines-pro-type-reinterpret-cast)
+        fillMSGHdr(&msgh, &iov, &cbuf, sizeof(cbuf), reinterpret_cast<char*>(&packet.at(0)), param.maxIncomingPacketSize, &remote);
+        while (true) {
+          try {
+            handleOnePacket(param);
+          }
+          catch (const std::bad_alloc& e) {
+            /* most exceptions are handled by handleOnePacket(), but we might be out of memory (std::bad_alloc)
+               in which case we DO NOT want to log (as it would trigger another memory allocation attempt
+               that might throw as well) but wait a bit (one millisecond) and then try to recover */
+            usleep(1000);
+          }
+        }
+      }
+      else {
+        auto callback = [&remote, &msgh, &iov, &packet, &handleOnePacket, initialBufferSize](int socket, FDMultiplexer::funcparam_t& funcparam) {
+          (void)socket;
+          const auto* param = boost::any_cast<const UDPStateParam*>(funcparam);
+          try {
+            remote.sin4.sin_family = param->cs->local.sin4.sin_family;
+            packet.resize(initialBufferSize);
+            /* used by HarvestDestinationAddress */
+            cmsgbuf_aligned cbuf;
+            // NOLINTNEXTLINE(cppcoreguidelines-pro-type-reinterpret-cast)
+            fillMSGHdr(&msgh, &iov, &cbuf, sizeof(cbuf), reinterpret_cast<char*>(&packet.at(0)), param->maxIncomingPacketSize, &remote);
+            handleOnePacket(*param);
+          }
+          catch (const std::bad_alloc& e) {
+            /* most exceptions are handled by handleOnePacket(), but we might be out of memory (std::bad_alloc)
+               in which case we DO NOT want to log (as it would trigger another memory allocation attempt
+               that might throw as well) but wait a bit (one millisecond) and then try to recover */
+            usleep(1000);
+          }
+        };
+        auto mplexer = std::unique_ptr<FDMultiplexer>(FDMultiplexer::getMultiplexerSilent(params.size()));
+        for (const auto& param : params) {
+          mplexer->addReadFD(param.socket, callback, &param);
+        }
+
+        timeval now{};
+        while (true) {
+          mplexer->run(&now, -1);
+        }
+      }
+    }
+  }
+  catch (const std::exception& e) {
+    SLOG(errlog("UDP client thread died because of exception: %s", e.what()),
+         dnsdist::logging::getTopLogger("udp-frontend")->error(Logr::Error, e.what(), "UDP client thread died because of exception"));
+  }
+  catch (const PDNSException& e) {
+    SLOG(errlog("UDP client thread died because of PowerDNS exception: %s", e.reason),
+         dnsdist::logging::getTopLogger("udp-frontend")->error(Logr::Error, e.reason, "UDP client thread died because of PowerDNS exception"));
+  }
+  catch (...) {
+    SLOG(errlog("UDP client thread died because of an exception: unknown"),
+         dnsdist::logging::getTopLogger("udp-frontend")->info(Logr::Error, "UDP client thread died because of an unknown exception"));
   }
 }
